reviews resolved
