+++ app/assets/stylesheets/variable.scss
@@ -1,22 +1,18 @@
 $red: #FF0000;
-$green: #94B26F;
-$light_green: #B6D491;
-$yellow: #F2DC87;
-$dark_brown: #332318;
-$light_brown: #4C2516;
 $white: #FFFFFF;
 $black: #000000;
 $bunker: #2E2F30;
 $cornflowerblue: #6495ED;
-$lightgray: #D3D3D3;
+$gray: #808080;
 $teal: #008080;
-$tomato: #ff634;
+$tomato: #ff6347;
 $whitesmoke1: #f5f5f5;
 $whitesmoke2: #EFEFEF;
 $darkred: #8B0000;
+$silver: #C0C0C0;
 $crimson: #DC143C;
-$hexgray: #cccccc;
-$hex1: #dddddd;
-$hex2: #f1f1f1;
-$gray40: #666666;
-$outrageousorange: #FF5E1F
+$darkgray: #666666;
+$outrageousorange: #FF5E1F;
+$light-gray: lighten($silver, 20);
+$hexgray1: lighten($silver, 30);
+$hexgray2: lighten($silver, 10);
