+++ app/views/clearance_mailer/change_password.html.erb
@@ -5,4 +5,4 @@
     edit_user_password_url(@user, token: @user.confirmation_token.html_safe) %>
 </p>
 
-<p><%= raw t('.closing') %></p>
+<p><%= raw t('clearance_mailer.change_password.closing_html') %></p>
