+++ app/views/clearance_mailer/change_password.text.erb
@@ -2,4 +2,4 @@
 
 <%= edit_user_password_url(@user, token: @user.confirmation_token.html_safe) %>
 
-<%= raw t(".closing") %>
+<%= raw t("clearance_mailer.change_password.closing_html") %>
