+++ app/views/sessions/_form.html.erb
@@ -2,6 +2,6 @@
 <%= f.input :email, placeholder: 'userdomain.com', label: t('shared.all.email') %>
 <%= f.input :password, label: t('shared.all.password')%>
 <%= f.button :submit, t('users.new.sign_in') %>
-<div class="other-links" style="margin: 10px 200px; width: 212px; height: 30px;">
+<div class="other-links">
   <%= link_to t('.forgot_password'), new_password_path %>
 </div>
