+++ app/views/user_mailer/sign_up_confirmation.text.erb
@@ -1,3 +1,3 @@
 <p><%= t('.opening') %></p>
 <%= confirm_email_url(token: @user.confirmation_token) %>
-<p><%= raw t('.closing') %></p>
+<p><%= t('user_mailer.sign_up_confirmation.closing_html') %></p>
