+++ config/routes.rb
@@ -1,6 +1,4 @@
 Rails.application.routes.draw do
-  post '/session', controller: :sessions, action: :create
-  delete '/sign_out', controller: :sessions, action: :destroy, as: 'log_out'
   get '/sign_up' => 'users#new', :as => 'register'
 
   get '/confirm_email/:token', to: 'email_confirmations#confirm', as: 'confirm_email'
