+++ spec/mailers/user_mailer_spec.rb
@@ -9,7 +9,7 @@ RSpec.describe UserMailer, type: :mailer do
     end
 
     it 'renders the receiver email' do
-      expect(mail.to).to eq(['user14@example.com'])
+      expect(mail.to).to eq(['user6@example.com'])
     end
   end
 end
